Move evaluator into src
